Move BadClientSilencer from example to package io.netty.handler.codec.http
